Extend the linau_conv_field interface.
Now there is a union inside the linau_conv_field structure.  It could
be a lcf_validate function like before or a lcf_match function which
returns an nvlist of matching fields.  The point is that it allows us to
detect and handle regex fields.

A regex field is a field the name of which is not predefined as it is
defined with a regex. A regex field is not "a0" while a regex field is
"a4" for example.