rc.d/sendmail: Fix the rcorder block
rcorder(8) requires the rcorder block to be an uninterrupted sequence of
REQUIRE, PROVIDE, BEFORE, and KEYWORD lines. Having a comment in between
REQUIRE and KEYWORD makes rcorder stop parsing the block when it reaches
the comment. Fix that by moving the comment out from the rcorder block.

MFC after:	3 days