rc.subr(8): Set procname to command by default to match documentation
The rc.subr(8) manual says that variable procname defaults
to the value of command. However, procname is never actually set to any
value in the run_rc_command().

This patch sets procname to the value of command if procname is unset
or empty. As a result it is going to be possible to use the procname
variable in *cmd functions (e.g., start_cmd).

Fixes:	64deb069a979 Bring this up to date with the latest NetBSD bits.
MFC after:	1 week
Differential Revision:	https://reviews.freebsd.org/D21450